requirement.txt
azure-functions
msal
requests
pandas
openpyxl
python-dotenv
sqlalchemy
pyodbc